refactoring-kata-master/src/main/java/com/crazyfingers/Inventory.java
package com.crazyfingers;

class Inventory {
    Item[] items;

    public Inventory(Item[] items) {
        this.items = items;
    }

    public void updateQuality() {
        for (int i = 0; i < items.length; i++) {
            if (!items[i].name.equals("FM-Pro4")
                    && !items[i].name.equals("FM-Tco4 LCV")) {
                if (items[i].quality > 0) {
                    if (items[i].quality > 50) {
                        items[i].quality = 50;
                    }
                    if (!items[i].name.equals("Fuel level sensor")&&(items[i].quality>=0)) {
                        items[i].quality = items[i].quality - 1;
                    }
                }
            } else {
                if (items[i].quality < 50) {
                    if (!items[i].name.equals("FM-Tco4 LCV")) {
                        items[i].quality = items[i].quality + 1;
                    }
                    if (items[i].name.equals("FM-Tco4 LCV")) {
                        if (items[i].sellIn <= 11 && items[i].sellIn > 6) {
                            if (items[i].quality < 50) {
                                items[i].quality = items[i].quality + 2;
                            }
                        }

                        if (items[i].sellIn <= 6) {
                            if (items[i].quality < 50) {
                                items[i].quality = items[i].quality + 3;
                            }
                        }
                        if (items[i].sellIn > 11) {
                            if (items[i].quality < 50) {
                                items[i].quality = items[i].quality - 1;
                            }
                        }
                    }

                }
            }

            if (!items[i].name.equals("Fuel level sensor")) {
                items[i].sellIn = items[i].sellIn - 1;
            }

            if (items[i].sellIn < 0 ) {
                if (!items[i].name.equals("FM-Pro4")) {
                    if (!items[i].name.equals("FM-Tco4 LCV")) {
                        if (items[i].quality > 0) {
                            if (!items[i].name.equals("Fuel level sensor")) {
                                items[i].quality = items[i].quality - 1;
                            }
                        }
                    } else {
                        items[i].quality = 0;
                    }
                } else {
                    if (items[i].quality < 50) {
                        items[i].quality = items[i].quality + 1;
                    }
                }
            }
        }
    }
}
